requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-admin-thumbnails==0.2.8
django-allauth==0.61.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-jalali-date==1.1.0
django-js-asset==2.2.0
django-rosetta==0.10.0
idna==3.6
jdatetime==4.1.1
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.2.0
polib==1.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.42.0
